+++ NEWS.md
@@ -1,5 +1,7 @@
 # rdocdump (development version)
 
+* Added support for remote repository references (GitHub, GitLab, Bitbucket) in `rdd_to_txt()` and `rdd_extract_code()`. This includes support for specific branches/tags/commits and packages in subdirectories, as well as direct URLs (to branches and/or folders within repositories). Therefore the `remotes` package to `Suggests` to handle remote downloads.
+
 # rdocdump 0.1.1 (2025-08-21)
 
 * Added `version` argument to `rdd_to_txt()` and `rdd_extract_code()` functions to specify the version of the package to download from CRAN for processing.